fix: refactored the code and added tests for critical path
