+++ package.json
@@ -4,19 +4,32 @@
   "description": "This Github Action allows developers to test and publish user transformations and libraries directly from their development repository",
   "main": "src/index.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
+    "test": "jest && codecov",
     "build": "ncc build --no-cache src/index.js  -o dist"
   },
   "author": "",
   "license": "ISC",
+  "jest": {
+    "coverageDirectory": "./coverage/",
+    "collectCoverage": true
+  },
   "dependencies": {
     "@actions/artifact": "^1.1.2",
     "@actions/core": "^1.10.0",
     "axios": "^1.6.4",
+    "codecov": "^3.8.3",
     "deep-object-diff": "^1.1.9",
     "lodash": "^4.17.21"
   },
   "devDependencies": {
+<<<<<<< HEAD
     "@vercel/ncc": "^0.36.1"
+=======
+    "@vercel/ncc": "^0.36.1",
+    "jest": "^29.7.0"
+  },
+  "engines": {
+    "npm": "^9.6.7"
+>>>>>>> 7333ed9 (fix: refactored the code and added tests for critical path)
   }
 }
