+++ src/testdata/emptymeta.json
@@ -1,0 +1,4 @@
+{
+  "transformations": [],
+  "libraries": []
+}
