+++ src/testdata/events.json
@@ -1,0 +1,39 @@
+[
+  {
+    "userId": "identified user id",
+    "anonymousId": "anon-id-new",
+    "messageId": "message-id-1",
+    "context": {
+      "ip": "14.5.67.21",
+      "library": {
+        "name": "http"
+      },
+      "traits": {
+        "address": {
+          "city": "Kolkata",
+          "country": "India",
+          "street": "330/8"
+        }
+      }
+    },
+    "timestamp": "2020-02-02T00:23:09.544Z"
+  },
+  {
+    "userId": "identified user id",
+    "anonymousId": "anon-id-new",
+    "messageId": "message-id-2",
+    "context": {
+      "ip": "14.5.67.21",
+      "library": {
+        "name": "http"
+      },
+      "traits": {}
+    },
+    "properties": {
+      "price": 20,
+      "revenue": 15,
+      "street": "330/8"
+    },
+    "timestamp": "2020-02-02T00:23:09.544Z"
+  }
+]
