+++ src/testdata/expected_diff.json
@@ -1,0 +1,10 @@
+{
+  "added": {},
+  "deleted": {},
+  "updated": {
+    "0": {
+      "revenue": 1,
+      "price": 1
+    }
+  }
+}
